Move Ad out of elements folder
