+++ packages/frontend/amp/components/Ad.tsx
